+++ Cargo.toml
@@ -12,6 +12,7 @@ members = [
 prisma_client = { path = "shared-libs/rust/prisma-client" }
 services = { path = "shared-libs/rust/services" }
 error_handler = { path = "shared-libs/rust/error-handler" }
+config = { path = "shared-libs/rust/config" }
 prisma-client-rust = { git = "https://github.com/Brendonovich/prisma-client-rust", default-features = false, features=["postgresql", "rspc", "migrations", "mocking"], rev = "13fdf89503de65156011b9a135d9d12fb589e8f8"}
 prisma-client-rust-cli = { git = "https://github.com/Brendonovich/prisma-client-rust", default-features = false, features=["postgresql", "rspc", "migrations", "mocking"], rev = "13fdf89503de65156011b9a135d9d12fb589e8f8"}
 redis = {version = "^0.22", features = ["aio", "ahash", "tokio-comp", "json"]}
