+++ apps/erudio/backend-api/src/helpers/consts.rs
@@ -1,0 +1,1 @@
+pub const SESSION_COOKIE_NAME: &str = "SessionId";
