+++ configs/dev/Config.ron
@@ -13,6 +13,7 @@ Config (
     salt_size: 32,
     secret_size: 32,
     db_url: "postgresql://postgres:password@localhost:5432/erudio?schema=public",
+    db_url_test: "postgresql://postgres:password@localhost:5432/erudio?schema=test",
     redis_url: "redis://127.0.0.1/",
     region_id: "TEST_REGION",
     api_port: 3001,
