+++ shared-libs/rust/services/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2021"
 prisma_client.workspace = true
 error_handler.workspace = true
 redis.workspace = true
+config.workspace = true
 log = "^0.4"
 serde = "^1.0"
 hex = "^0.4"
@@ -28,3 +29,4 @@ redis-test = { version = "0.1.1", features=["aio"] }
 once_cell = "^1.16"
 eyre = "^0.6"
 uuid = { version = "^1.2", features=["v4"]}
+ron = "^0.8"
